Update SPI.h
I need to change SPI settings in the DmaSpi library. As all data is private to SPISettings, I propose an extra method in SPIClass for that purpose (which is a friend of SPISettings and thus can access the internals).

The DmaSpi handles one or more slaves between a single pair of beginTransaction() and endTransaction() calls, by letting a chip select class handle the cs signal and apply SPI settings.