Fix Fifo issue.
There is an issue where with some SPI devices (ILI9341_t3) when they
finish a transaction, the receive fifo may not be empty, which causes
simple SPI.transfer to return the wrong data.

This fix is to empty the fifos each time we begin a transaction.
More details in the thread:

https://forum.pjrc.com/threads/26305-Highly-optimized-ILI9341-%28320x240-TFT-color-display%29-library?p=71210&viewfull=1#post71210